refactor: extract pipeline/git.rs
Move git-source handling out of pipeline/mod.rs into a focused
submodule: clone URL construction (github_clone_url, gitlab_clone_url),
authentication injection (inject_basic_auth, percent_encode_userinfo),
shallow-clone-to-tempdir (clone_to_tempdir), and the InstallSource
dispatch entrypoints (install_from_source, fetch_ssot,
install_from_git_url). Inline tests move with the functions.

Public API is preserved via 'pub use git::{...}' in mod.rs.

Refs #180
